requirements.txt
tensorflow-gpu==1.15.0
sentencepiece==0.1.85
pickle
tqdm==4.41.0
networkx==2.4

